Add topology HTML synchronized state
