+++ README.md
@@ -133,7 +133,7 @@ For a single chart, `WithXAxisBounds(minimum, maximum)` and `WithYAxisBounds(min
 
 Use `WithXAxisVisible(false)` or `WithYAxisVisible(false)` when an infographic-style chart should keep one axis worth of labels while removing the other axis line, ticks, and title. Use `WithAxisLines(false)` when tick labels and titles should remain but the axis rules and zero-lines should disappear. For example, horizontal scorecards often keep category labels on the left and hide the numeric value axis plus the remaining axis rule.
 
-Topology diagrams live in `ChartForgeX.Topology` because they are diagram models, not dashboard pages or data collectors. `TopologyChart` owns reusable groups, nodes, edges, deterministic layout hints, validation, SVG rendering, PNG rendering, and a tiny HTML wrapper that embeds the SVG in a neutral `<div>`. Complete topology HTML pages include lightweight pointer/focus hover, click selection, and keyboard selection by default. They dispatch `cfx-topology-hover`, `cfx-topology-hover-clear`, `cfx-topology-select`, `cfx-topology-clear`, and `cfx-topology-navigate` with selected ids, kind/status, metadata, metrics, route diagnostics, geographic route-arc diagnostics, endpoint context, and related node/edge/group ids; hosts can also dispatch `cfx-topology-set-selection` or `cfx-topology-clear-selection` to control state. Arrow keys cycle focus through related topology elements. Set `EnableHtmlViewportControls = true` to add opt-in zoom/pan/reset controls with `cfx-topology-viewport` host events and host-driven `cfx-topology-set-viewport` / `cfx-topology-reset-viewport` support. Set `EnableHtmlExportControls = true` to add opt-in SVG/PNG export buttons with `cfx-topology-export` host events. Set `EnableHtmlInteractions = false` when HtmlForgeX or another host wants to own all behavior. HtmlForgeX can later host the rendered SVG in cards, filters, tabs, and inspector panels, while TestimoX or another product can map its own collected data into the model.
+Topology diagrams live in `ChartForgeX.Topology` because they are diagram models, not dashboard pages or data collectors. `TopologyChart` owns reusable groups, nodes, edges, deterministic layout hints, validation, SVG rendering, PNG rendering, and a tiny HTML wrapper that embeds the SVG in a neutral `<div>`. Complete topology HTML pages include lightweight pointer/focus hover, click selection, and keyboard selection by default. They dispatch `cfx-topology-hover`, `cfx-topology-hover-clear`, `cfx-topology-select`, `cfx-topology-clear`, and `cfx-topology-navigate` with selected ids, kind/status, metadata, metrics, route diagnostics, geographic route-arc diagnostics, endpoint context, and related node/edge/group ids; hosts can also dispatch `cfx-topology-set-selection` or `cfx-topology-clear-selection` to control state. Arrow keys cycle focus through related topology elements. Set `EnableHtmlViewportControls = true` to add opt-in zoom/pan/reset controls with `cfx-topology-viewport` host events and host-driven `cfx-topology-set-viewport` / `cfx-topology-reset-viewport` support. Set `EnableHtmlExportControls = true` to add opt-in SVG/PNG export buttons with `cfx-topology-export` host events. Set `EnableHtmlSynchronizedState = true` with `HtmlSyncGroupName` to mirror selection clears and viewport state across topology wrappers on the same page. Set `EnableHtmlInteractions = false` when HtmlForgeX or another host wants to own all behavior. HtmlForgeX can later host the rendered SVG in cards, filters, tabs, and inspector panels, while TestimoX or another product can map its own collected data into the model.
 
 ```csharp
 using ChartForgeX.Topology;
