+++ docs/topology-next-plan.md
@@ -120,11 +120,11 @@ Suggested PR after ChartForgeX routing/layout: `Add HtmlForgeX topology host pan
 
 ### 6. Deeper Browser-Side Interactions
 
-Topology HTML now provides pointer/focus hover, click selection, and keyboard selection with richer payloads. Hover highlights related nodes, edges, and groups, then dispatches `cfx-topology-hover` / `cfx-topology-hover-clear`. Arrow keys cycle focus through connected topology elements and dispatch `cfx-topology-navigate` with from/to details. A selected element dispatches id, kind, status, metadata, metrics, node/group/edge context, route diagnostics, geographic route-arc diagnostics, and related node/edge/group ids so host dashboards can drive inspector panels without reparsing the SVG. Hosts can also dispatch `cfx-topology-set-selection` and `cfx-topology-clear-selection` on the wrapper to synchronize external panels with the embedded topology. Large complete topology pages can opt into lightweight zoom, pan, wheel-zoom, reset, host-driven viewport state, and SVG/PNG export controls through `EnableHtmlViewportControls` and `EnableHtmlExportControls`. The general interactive chart wrapper still has brush, synchronized charts, and richer controls that topology does not yet reuse.
+Topology HTML now provides pointer/focus hover, click selection, and keyboard selection with richer payloads. Hover highlights related nodes, edges, and groups, then dispatches `cfx-topology-hover` / `cfx-topology-hover-clear`. Arrow keys cycle focus through connected topology elements and dispatch `cfx-topology-navigate` with from/to details. A selected element dispatches id, kind, status, metadata, metrics, node/group/edge context, route diagnostics, geographic route-arc diagnostics, and related node/edge/group ids so host dashboards can drive inspector panels without reparsing the SVG. Hosts can also dispatch `cfx-topology-set-selection` and `cfx-topology-clear-selection` on the wrapper to synchronize external panels with the embedded topology. Large complete topology pages can opt into lightweight zoom, pan, wheel-zoom, reset, host-driven viewport state, SVG/PNG export controls, and same-page selection/viewport synchronization through `EnableHtmlViewportControls`, `EnableHtmlExportControls`, and `EnableHtmlSynchronizedState`. The general interactive chart wrapper still has brush and richer controls that topology does not yet reuse.
 
 Next work:
 
-- unify topology HTML with `ChartForgeX.Interactivity.Html` where practical
+- unify topology HTML with `ChartForgeX.Interactivity.Html` where practical - topology-local sync/export/viewport controls are in place; shared adapter reuse remains open
 - add optional zoom/pan controls for large topology diagrams - lightweight topology-local controls are in place; adapter-level reuse remains open
 - dispatch richer selection details: node/edge/group id, kind, status, metrics, metadata, source/target, connected edges - done for the default complete-page selection event
 - support hover highlighting, keyboard navigation, and host-controlled selection state - pointer/focus hover, keyboard activation, arrow-key graph navigation, and host selection events are in place
