+++ docs/topology.md
@@ -111,7 +111,7 @@ var metricSvg = chart.ToSvg(new TopologyRenderOptions {
 });
 ```
 
-`TopologySvgRenderer` outputs a complete standalone SVG with `viewBox`, `defs`, scoped CSS, groups below edges, edge labels above edges, nodes above labels, status badges, optional legends, and accessibility metadata. `TopologyPngRenderer` draws the same model through ChartForgeX's dependency-free raster canvas for report exports. `TopologyHtmlRenderer` wraps the generated SVG in a neutral `.cfx-topology-wrapper` div with chart metadata and, by default for complete pages, a tiny interaction script that marks clicked groups, nodes, or edges and dispatches `cfx-topology-select` / `cfx-topology-clear` events. Selection details include the selected id, kind, status, metadata, metrics, group/node/edge context, route diagnostics, geographic route-arc diagnostics, and related node/edge/group ids so HtmlForgeX or TestimoX can populate inspector panels without reparsing SVG markup. Selectable topology elements are keyboard focusable, Enter/Space activates selection, Escape clears selection, pointer/focus hover highlights related elements and dispatches `cfx-topology-hover` / `cfx-topology-hover-clear`, arrow keys cycle focus through related topology elements and dispatch `cfx-topology-navigate`, and hosts can dispatch `cfx-topology-set-selection` or `cfx-topology-clear-selection` on the wrapper to control state. Set `TopologyRenderOptions.EnableHtmlViewportControls = true` to add opt-in zoom, pan, wheel-zoom, reset, `cfx-topology-viewport`, `cfx-topology-set-viewport`, and `cfx-topology-reset-viewport` support for large diagrams. Set `TopologyRenderOptions.EnableHtmlExportControls = true` to add opt-in SVG/PNG export buttons and `cfx-topology-export` events. Set `TopologyRenderOptions.EnableHtmlInteractions = false` when a host wants a purely static page or will provide all behavior itself.
+`TopologySvgRenderer` outputs a complete standalone SVG with `viewBox`, `defs`, scoped CSS, groups below edges, edge labels above edges, nodes above labels, status badges, optional legends, and accessibility metadata. `TopologyPngRenderer` draws the same model through ChartForgeX's dependency-free raster canvas for report exports. `TopologyHtmlRenderer` wraps the generated SVG in a neutral `.cfx-topology-wrapper` div with chart metadata and, by default for complete pages, a tiny interaction script that marks clicked groups, nodes, or edges and dispatches `cfx-topology-select` / `cfx-topology-clear` events. Selection details include the selected id, kind, status, metadata, metrics, group/node/edge context, route diagnostics, geographic route-arc diagnostics, and related node/edge/group ids so HtmlForgeX or TestimoX can populate inspector panels without reparsing SVG markup. Selectable topology elements are keyboard focusable, Enter/Space activates selection, Escape clears selection, pointer/focus hover highlights related elements and dispatches `cfx-topology-hover` / `cfx-topology-hover-clear`, arrow keys cycle focus through related topology elements and dispatch `cfx-topology-navigate`, and hosts can dispatch `cfx-topology-set-selection` or `cfx-topology-clear-selection` on the wrapper to control state. Set `TopologyRenderOptions.EnableHtmlViewportControls = true` to add opt-in zoom, pan, wheel-zoom, reset, `cfx-topology-viewport`, `cfx-topology-set-viewport`, and `cfx-topology-reset-viewport` support for large diagrams. Set `TopologyRenderOptions.EnableHtmlExportControls = true` to add opt-in SVG/PNG export buttons and `cfx-topology-export` events. Set `TopologyRenderOptions.EnableHtmlSynchronizedState = true` with `HtmlSyncGroupName` to mirror selection clears and viewport state across same-page topology wrappers using `cfx-topology-sync` / `cfx-topology-apply-sync`. Set `TopologyRenderOptions.EnableHtmlInteractions = false` when a host wants a purely static page or will provide all behavior itself.
 
 `TopologyRenderOptions` also supports reusable render presets and node display modes:
